requirements.txt
Flask==2.1.2
psycopg2-binary==2.9.1
Werkzeug==2.1.2
